+++ pkg/coordinator/logger/logreader.go
@@ -3,6 +3,6 @@ package logger
 import "github.com/ethpandaops/assertoor/pkg/coordinator/db"
 
 type LogReader interface {
-	GetLogEntries(from, limit int) []*db.TaskLog
-	GetLogEntryCount() int
+	GetLogEntries(from, limit uint64) []*db.TaskLog
+	GetLogEntryCount() uint64
 }
