+++ pkg/coordinator/types/coordinator.go
@@ -23,7 +23,7 @@ type Coordinator interface {
 
 	GetTestByRunID(runID uint64) Test
 	GetTestQueue() []Test
-	GetTestHistory(testID string, firstRunID uint64, offset uint64, limit uint64) ([]Test, int)
+	GetTestHistory(testID string, firstRunID uint64, offset uint64, limit uint64) ([]Test, uint64)
 	ScheduleTest(descriptor TestDescriptor, configOverrides map[string]any, allowDuplicate bool) (TestRunner, error)
 	DeleteTestRun(runID uint64) error
 }
